Conflicts:
	spring-xd-dirt/src/main/java/org/springframework/xd/dirt/server/AdminServerApplication.java
	spring-xd-dirt/src/main/java/org/springframework/xd/dirt/server/LauncherApplication.java

+++ spring-xd-shell/src/test/java/org/springframework/xd/shell/AbstractShellIntegrationTest.java
@@ -38,6 +38,7 @@ import org.springframework.shell.core.CommandResult;
 import org.springframework.shell.core.JLineShellComponent;
 import org.springframework.xd.dirt.container.store.RedisRuntimeContainerInfoRepository;
 import org.springframework.xd.dirt.server.SingleNodeApplication;
+import org.springframework.xd.test.RandomConfigurationSupport;
 import org.springframework.xd.test.redis.RedisTestSupport;
 
 /**
@@ -54,7 +55,7 @@ import org.springframework.xd.test.redis.RedisTestSupport;
  * @author David Turanski
  * 
  */
-public abstract class AbstractShellIntegrationTest {
+public abstract class AbstractShellIntegrationTest extends RandomConfigurationSupport {
 
 	/**
 	 * Where test module definition assets reside, relative to this project cwd.
@@ -67,6 +68,8 @@ public abstract class AbstractShellIntegrationTest {
 	private static final File TEST_MODULES_TARGET = new File("../modules/");
 
 	protected static final String DEFAULT_METRIC_NAME = "bar";
+
+	public static boolean SHUTDOWN_AFTER_RUN = true;
 
 	@ClassRule
 	public static RedisTestSupport redisAvailableRule = new RedisTestSupport();
@@ -81,30 +84,35 @@ public abstract class AbstractShellIntegrationTest {
 
 	private static RedisRuntimeContainerInfoRepository runtimeInformationRepository;
 
-
 	@BeforeClass
-	public static void startUp() throws InterruptedException, IOException {
+	public static synchronized void startUp() throws InterruptedException, IOException {
+		if (application == null) {
+			application = new SingleNodeApplication().run("--transport", "local",
+					"--analytics", "redis",
+					"--store", "redis"
+					);
+			Bootstrap bootstrap = new Bootstrap(new String[] { "--port",
+				RandomConfigurationSupport.getAdminServerPort() });
+			shell = bootstrap.getJLineShellComponent();
 
-		application = new SingleNodeApplication().run("--transport", "local",
-				"--analytics", "redis",
-				"--store", "redis"
-				);
-		Bootstrap bootstrap = new Bootstrap();
-		shell = bootstrap.getJLineShellComponent();
-
-		runtimeInformationRepository = application.getContainerContext().getBean(
-				RedisRuntimeContainerInfoRepository.class);
-
+			runtimeInformationRepository = application.getContainerContext().getBean(
+					RedisRuntimeContainerInfoRepository.class);
+		}
+		if (!shell.isRunning()) {
+			shell.start();
+		}
 	}
 
 	@AfterClass
 	public static void shutdown() {
-		runtimeInformationRepository.delete(application.getContainerContext().getId());
-		logger.info("Stopping XD Shell");
-		shell.stop();
-		if (application != null) {
-			logger.info("Stopping Single Node Server");
-			application.close();
+		if (SHUTDOWN_AFTER_RUN) {
+			runtimeInformationRepository.delete(application.getContainerContext().getId());
+			logger.info("Stopping XD Shell");
+			shell.stop();
+			if (application != null) {
+				logger.info("Stopping Single Node Server");
+				application.close();
+			}
 		}
 	}
